Add abstract interface LibraryDB definitions
The core of the database module will be this interface class,
LibraryDB. It is an abstract class so that the users do not depend on
the actual implementation of the Database, be it SQL or another
database manager.

Such class will be implemented (inherited) by another subclass
specifying the database manager system and implementing the virtual
methods.